Booster flag to disable simplification before fall-back request (#3825)
Logs are showing that the simplification performed before fall-back
execute requests is often not productive. This PR provides a server flag
to switch it off and save the overhead.

OTOH, the simplification may prune spurious branches, saving a
`Branching` re-execution in `kore-rpc` in some cases. Therefore the
default behaviour is still to simplify before fallback requests.