Do not output attribute information when unparsing `TermLike` (#3823)
When pretty-printing a `TermLike`, the term attributes are serialized
into text and in-lined into the unparsed Kore term, leading to results
like this one:
```
    Could not infer the equation requirement:
        /* Spa */
        \equals{SortBool{}, _}(
            /* T Fn D Spa */
            Lbl'Unds-LT-'Int'Unds'{}(
                /* T Fn D Spa */
                Lblf'LParUndsRParUnds'TEST'Unds'Int'Unds'Int{}(
                    /* T Fn D Sfa */ ConfigVarI:SortInt{}
                ),
                /* T Fn D Sfa Cl */ \dv{SortInt{}}("0")
            ),
            /* T Fn D Sfa Cl */ \dv{SortBool{}}("true")
        )
```

I'd argue that we do not care as much about the exact attributes of
`TermLike`s anymore today. Therefore I propose to remove the attributes,
leading to a much cleaner log output:

```
    Could not infer the equation requirement:
        \equals{SortBool{}, _}(
            Lbl'Unds-LT-'Int'Unds'{}(
                Lblf'LParUndsRParUnds'TEST'Unds'Int'Unds'Int{}(
                    ConfigVarI:SortInt{}
                ),
                \dv{SortInt{}}("0")
            ),
            \dv{SortBool{}}("true")
        )
```

The integration tests fail, but I will fix them once we have decided if
we want this merged.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>