Add details to evaluation aborts that cause warnings
For warnings during evaluation/simplification, the warning message
advises users to use a certain log context. This PR changes the
recommended log context to `*abort*detail`, for which informative
details have been added in all call sites of `logWarn`.

At the same time, the warning about too many iterations was duplicated
in the `Rewrite` module, and was therefore downgraded to `Simplify`
level there.