Reuse kompiled `contracts.k` and `foundry.k` in Kontrol performance test (#3845)
This PR attempts to reduce the influence of the K compiler processes on
the Kontrol performance test. To achieve that, we compile everything
in-advance, ensuring that there will be no calls to `kompile` when
running the actual proof test:

- run the single `test_foundy_kompile` test, which is essentially just a
wrapper for the `foundry` pytest
[fixture](https://github.com/runtimeverification/kontrol/blob/2fba281b0a73d76c1dba37f74220051901d04606/src/tests/integration/conftest.py#L53)
that generates and compiles the `contracts.k` and `foundry.k` files.
This produces the `foundry/out/kompiled` directory
- when running the actual proof tests, pass the generated `foundry`
directory to them explicitly. This makes sure that the pytest fixture
will not be forced to recompile stuff in every `pytest-xdist` worker.
- Re-using the `foundry` directory has a side-effect of also reusing the
`foundry/out/proofs` directory. To make sure that we actually run the
proofs in `master_shell`, and not re-use the ones produced by
`feature_shell`, we delete them between the steps.

Some things I've noticed that may be worth accounting for, but not in
this PR:
- the `server`
[fixture](https://github.com/runtimeverification/kontrol/blob/7f45d5672a59f094e2d067e41923e037544b86f1/src/tests/integration/conftest.py#L35)
is module-scoped. That means that all tests in a certain test module
will reuse the same server instance, individually in each test worker.
- the corollary of the above is that the proofs that are run later will
use a more "tired" server, since the server state will bloat due to a
bunch of duplicate (and likely empty) modules. It is not clear if this
is a performance issue, but we have certainly not considered it earlier.
- another thing is that the decision on which test to skip is taken
dynamically at the start of the test, and the tests, even if skipped,
need the fixtures to be ready for them. In most cases, however, the
server connection will be reused by other tests in the same module.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>