Fake `rule-id` with the rule label when rules don't have unique ids (#3916)
Fixes https://github.com/runtimeverification/haskell-backend/issues/3913

- in Booster, we change the representation of rules to enforce then to
have a `UniqueId`. When internalizing a rule, we see that it does not
have one, we use the rule label. If that is not there to, we use
`Unknown`.
- In Kore, we do not change the representation of rules, but rather
apply the same trick to the results, i.e. we fall back to the label or
"Unknown"

this approach is a workaround to make rules that are dynamically
generated by `pyk` visible.

The long-term solution to this problem would be to enforce all rules
produced by the K Frontrend/`pyk` to have unique ids.

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>
Co-authored-by: Jost Berthold <jost.berthold@gmail.com>