Set up K-style versioning (#3924)
This PR introduces a release versioning scheme that's a carbon copy of
the setup used in the LLVM backend, but with development happening on
the `master` branch and versioning information being carried on the
`release` branch.

The code is well-tested on the LLVM backend side, and I've tested the
various versioning scripts locally, but it's difficult to explicitly
test the end-end integration here without just merging the PR. Any
remaining bugs or issues will be easy to iron out once we've seen a run
happen here.