Remove cabal cache form CI (#3932)
We noticed that we are trying to cache the cabal store, however, Nix
already does this for us, as we have seen from the fact that cache has
been broken but builds run just as fast. Therefore, this caching should
be removed entirely.

This also fixes an issue where the cache key on master was not being
built correctly, because we forgot to set the `ghc_version` ENV
variable. THis also meant we were building with a different version of
GHC on master push vs the feature branches.