Attempt no. 2 at fixing rule-substitution (#3939)
Fixes #3912?

There were three further issues that I identified:
* When we have a `VarX` in both the rule and configuration, internally
they are distinguished by a different constructor. However, when
externalising, this tag was lost. This PR prepends "Rule" to any rule
variables and "Eq" to any Equation vars, leaving the original name for
all config vars.
* After match/unification, there is some further processing before the
substitution is applied. We are now capturing the simplified/normalised
substitution just before it's applied and emitted in a debug message
(via `debugCreatedSubstitution`). This should ensure that the
substitution is the correct one.
* The substitution being returned contained both the mappings from rule
variables to subterms of the configuration, as well as learned
predicates about equalities between configuration variables and values.
The latter are emitted in the `rule-predicate` field and should not be
present in `rule-substitution`.