Add hook for KRYPTO.keccak256raw (#3966)
There are a few more of these `KRYPTO.&_raw` functions declared with
hook without implementation. We could add them now, or as needed later.

The `*raw` family of (hooked) functions differs from the non-`raw`
functions in that they return their result as a (byte) string of
characters instead of encoding it as a sequence of hexa-decimal digits.

Example:
* when `keccak256(X)` returns `"deadbeef"`,
* then `keccak256raw(X)` returns `"\xde\xad\xbe\xef"`, or
`"\222\173\190\239"` in the notation used by Haskell

---------

Co-authored-by: github-actions <github-actions@github.com>