Consider known predicates in Booster's simplifier API (#3953)
The `runEquationT` function now takes a `Set Predicate`, which are the
"predicates known to be true" and are populated into the initial
`EquationState`. Their origin is intended to be the rewriting path
condition.

There's now a possibility to supply prior knowledge to the following API
functions in `Booster.Pattern.Rewrite`:
- `evaluateTerm`
- `simplifyConstraint`

There's also a light change in `evaluatePattern`, which is now passes
the pattern's constraints as an argument to `runEuationT` rather then
using `pushConstraint`.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>