Recover from `TooManyIterations` when simplifying during rewriting (#3989)
Booster's rewriting algorithm, upon getting stuck (or aborting due to
`RuleApplicationUnclear`), will try so simplify the configuration once
and retry applying rules. When we cannot not finish simplifying the term
due to a `TooManyIterations`, we throw the partial result away and fail.

This PR handles the `TooManyIterations` exception and attempts applying
rewrite rules to the partially simplified term.