Update `kore-rpc-client` to use `tar-0.6.3` (#3996)
With https://github.com/runtimeverification/k/pull/4548, the bug reports
will be generated in GNU format of tar files instead of the default
POSIX.1-2001 pax format. This is necessary because:
- the filenames in the bug reports are often longer than the
format-portable restriction of 256 characters
- the Haskell [library](https://hackage.haskell.org/package/tar-0.6.3.0)
we use to work with tar does not support the pax format. However, it
does support the GNU format which also allows long names.

With this change, we get to process bug reports with long names.