Booster: Avoid var. names from input substitution when creating new ones
When rewriting with rules that have existentials on the RHS, booster
needs to rename these existential variables to avoid collisions.

However, the input is pre-processed to recognise predicates of the
form `X ==K expr` and apply them (after checking/vetting) as a
_substitution_ to the configuration and (other) path conditions.

As a result, the variable names in the substitution are not present in
the pattern any more when rewriting. The substitution is added back
(as additional information) in the output, so subsequent
simplification steps will apply it again, falsely, if variable names
from the substitution are reused during rewriting.

As a solution, we supply the "forbidden variable names" to the
rewriter. Alternatively, we could stop including the applied
substitution in the output, but this would create a disparity between
`booster` and `kore-rpc`.