EXPERIMENT Only send unevaluated functions to LLVM
This change steps away from maximising the terms we send to the LLVM
backend for evaluation. It addresses a particular constellation that
has been observed in downstream semantics/proofs:
* a configuration containing large data structures in cells that remain unchanged
* a long-running sequence of rewrites that keep inserting concrete function applications into cells required for the next rewrite step,
* and therefore require frequent configuration simplifications

Each time the configuration is simplified, the large unchanged data
structure is sent to the LLVM backend, which incurs considerable
overhead.

On the flip side, we will send each function that needs to be
evaluated individually. However, rewrites or equations will usually
not create more than a handful of new unevaluated function calls at
one time, and we will still send _nested_ unevaluated expressions
together.