Pass `--ignore-environment` to nix dev shells in performance scripts (#4041)
Due to recent packaging changes to the blockchain plugin, the KEVM
performance script started to fail as it was using libraries provided by
brew and not nix on my local machine. The fix is to call `nix develop`
with `--ignore-environment`, which purges the path of any non-nix paths.
In general it is good practice to do this in any dev shells we call in
one of these scripts, as there is less likelihood of similar issues
arising.