Handle inconsistent ground truth and SMT unknowns when checking `ensures` (#4063)
When checking `ensures` conditions of rewrite rules with the SMT solver,
we must mark rewrite as trivial if the ground truth is inconsistent. If
the SMT solver returns unknown, we must abort rewriting.

Previously, we were swallowing both of there cases and finalizing the
rewrite step successfully. This behavior of Booster went undetected for
a long time since we would usually abort rewriting or detect a vacuous
state at the next step, resulting in wasted work but no unsoundness.

We also tweak the return type of `checkPredicates` to convey addition
information why the result is unknown. This will be useful when we start
tolerating SMT unknowns and branching on that.