Assume the rule remainder condition when checking ensures (#4071)
When Booster detects an indeterminate `requires`, it adds it to the
constraints on the rewritten `Pattern`. However, when we check the
`ensures` clause, we only pass the the constraints on the old pattern as
known truth. This PR corrects that by adding the uncertain conditions as
additional known truth, as the `ensures` check should be performed on
the rewritten pattern, not the old one.