HOTFIX Make some informative warning logs optional, comment about implies
The `logMessage'` function, which emits log messages _unconditionally_,
was used for logging some exception conditions in booster. However, these warnings
should not appear by default in user-facing output when booster is called through
the proxy and just falls back to using kore-rpc.

The changes do not remove the log messages, they are now emitted at context `[proxy]`,
which is a common and recommended log setting to see what the server is up to.

Besides changing the `logMessage'` calls, a comment is added to point out where
the incompleteness of the `implies` endpoint is rooted in the code.