Hotfix sortk equality issue (#4074)
Several errors were reported on slack which involve an erroneous
injection `inj{SortK, SortKItem}(_)`.

The reason was within the internalisation of `\equals` by booster:
`equals(A:SortK, B:SortK)` was internalised to `inj{SortK,SortKItem}(A)
~> .K ==K inj{SortK,SortKItem}(B) ~> .K` instead of simply `A ==K B`.

In the investigated case, the problem surfaced when a fallback returned
`#Not (#Equals (_:SortK, _: SortK))`, which then caused a failure when
re-internalising the result to do further execution steps. The added
integration tests do not fully reproduce this crash but cause ill-sorted
output which will be detected if the error is reintroduced.