HOTFIX abort rewrite when SMT solver times out (#4078)
When the SMT solver timed out while checking `requires` clauses of a
rule, the unclear conditions were returned in the same way as conditions
that were known to be indeterminate. This created bogus branches in
proofs when the solver had a problem to decide a condition.

On such timeouts, and on inconsistent ground truths, booster now aborts
the entire rewrite. This might lead to increased spurious aborts in
complex proofs, but is probably better than having to prune the bogus
branches from the client side. The fall-back to legacy kore was able to
prune the bogus branch easily in the case that was investigated.