4100 Abort booster rewrite when matching subject variables
As demonstrated in #4100, booster should not report `stuck` when attempting to rewrite a term with variables that are required in a rule match.
Specifically, a subject variable may lead to branching with the rule content matches ending up as path conditions in the different branches.

Previously, booster would discard such cases as match failures and report `stuck` since no rule was applied.
This PR changes the behaviour to aborting the rewrite when a subject variable needs to be matched by anything but a rule variable.