4100 booster abort rewrite on subject vars (#4101)
Fixes #4100 

During rewriting, booster was reporting the rule match as "failed" when
trying to match a non-variable in the rule to a variable in the subject
(introduced by a refactoring mid 2024).
This produces unsound results because the subject variable could be
constrained to have a value that would match the rule term, enabling the
application of this rule.
The result in `booster` should therefore be `aborted` as long as we
don't allow for introducing constraints like this to the rule
application.

The new code adds the subject variable to a set of "indeterminate" match
pairs. Code has been kept separate so we can later extend the branching
mechanism by the collected constraints on subject variables.