HOTFIX invalidate cache when modifying known predicates
The change in #3953 introduced additional arguments to evaluation
functions which allow callers to supply some "known true" predicates
for the simplification and evaluation. However, doing so means that
the cache will get populated with associations that might only be true
if this known truth does not change.

The case in point was a predicate being cached as "true" because of an
earlier evaluation, and then _removed_ from the path condition.

This change removes returning the cache from any inteface functions
that take known truth arguments, and runs internal computations that
modify the predicates with a fresh empty cache.