Update nix flake inputs structure and add dependency update workflow for `rv-nix-tools` (#4109)
This pull request adapts the nix flake inputs structure regarding
`nixpkgs` and `rv-nix-tools`. In particular, `nixpkgs` is now specified
for inputs and transitive inputs by following an explicitly specified
revision of `rv-nix-tools`. More thorough reasoning for this change is
outlined in documentation added by this PR
runtimeverification/rv-nix-tools#14. This change is done to make
`nixpkgs` properly updatable by CI.

In addition, this pull requests adapts the `update-deps.yml` workflow to
also update `rv-nix-tools` that is used to implicitly specify the
revision of `nixpkgs`.

---

###### Review checklist

The author performs the actions on the checklist. The reviewer evaluates
the work and checks the boxes as they are completed.

- [x] **Summary.** Write a summary of the changes. Explain what you did
to fix the issue, and why you did it. Present the changes in a logical
order. Instead of writing a summary in the pull request, you may push a
clean Git history.
- [ ] **Documentation.** Write documentation for new functions. Update
documentation for functions that changed, or complete documentation
where it is missing.
- [ ] **Tests.** Write unit tests for every change. Write the unit tests
that were missing before the changes. Include any examples from the
reported issue as integration tests.
- [ ] **Clean up.** The changes are already clean. Clean up anything
near the changes that you noticed while working. This does not mean only
spatially near the changes, but logically near: any code that interacts
with the changes!

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>