Fix LIST.in hook for non-constructor-like list elements
The previous list hook implementation `LIST.in` was testing membership
based on _syntactic_ equality. This is incorrect if the elements of
the list are not constructor-like (like variables or unevaluated function calls),
`False` cannot be returned in those cases. When all known list elements are
constructor-like, syntactic equality is sufficient.
A unit test was added for this case.