HOTFIX free after simplify bool, call munmap after each request (#4115)
Fixes a problem that surfaced in the `kontrol` upgrade (symptom: server
crashes unexpectedly after some iterations).
The LLVM backend was changed to not call `munmap` within
`free_all_kore_mem` any more, which makes the `mmap` calls to initialise
thread-local heap address space fail after a number of requests.

This is a hot-fix which calls `reset_munmap_all_arenas`, a function not
currently included in the LLVM backend C API.
The function will be included in the API once
https://github.com/runtimeverification/llvm-backend/pull/1210 is merged,
then the header file and the function name should be aligned with the
API again.

The backend is _not going to work with older LLVM backends_ because [the
function that needs to be called is new as of June
7](https://github.com/runtimeverification/llvm-backend/commit/facee2b4a2ed0b8d7797b53300a10170f96fdd91).