Only check top-level symbol when deciding to skip sort predicate rules
The previous fix checked `isEvaluated` on the term in a sort predicate
before applying rules. This is correct but leads to too many
fall-backs which slow down symbolic execution when sort predicates are
used in rewrite side conditions. The only case that matters is when an
_unevaluated function_ is at the _head_ of the term in the sort
predicate. This is now matched directly.